editor extracted from renderer and include cleanup

+++ include/block/Editor.hpp
@@ -1,8 +1,7 @@
 #pragma once
 
+#include <SFML/Window/Event.hpp>
 #include <string>
-
-#include <SFML/Graphics/RenderWindow.hpp>
 
 #include "Block.hpp"
 
@@ -34,6 +33,7 @@ inline bool isCoordConType(const ObjAtCoordVar& ref) {
     }
 }
 
+// Editor is responsible for storing and updating the state of the editor gui
 class Editor {
   private:
     ObjAtCoordVar whatIsAtCoord(const sf::Vector2i& coord) const;
@@ -62,7 +62,7 @@ class Editor {
     bool                          conStartLegal;
     bool                          conEndLegal;
 
-    sf::Vector2i snapToGrid(const sf::Vector2f& pos) const;
-    void         event(const sf::Event& event, const sf::Vector2f& mousePos);
-    void         frame(const sf::Vector2f& mousePos);
+    sf::Vector2i snapToGrid(const sf::Vector2f& pos) const; // should be in block
+    void         event(const sf::Event& event, const sf::Vector2i& mousePos);
+    void         frame(const sf::Vector2i& mousePos);
 };
